Add symbiosis toast (#23)
* feat(ui): add sonner dependency for improved toast notifications

Added the sonner library as a dependency to enhance toast notification functionality in the UI package. This addition will provide a more robust and customizable solution for displaying notifications to users, improving the overall user experience and interface feedback mechanisms.

* feat(ui): add Toaster component and useToast hook

Introduces a new Toaster component and useToast hook to provide a consistent and customizable toast notification system across the application. The Toaster component wraps the sonner library's Toaster, while the useToast hook offers a simplified API for creating and managing toasts with various configurations and variants.

* feat: Introduce storybook stories to showcase Toaster and useToast

* feat(ui): enhance toast functionality and add new icons

This commit introduces several improvements to the toast system and adds new icons for various toast types. The changes include:

1. New SVG icons for loader and different toast types.
2. Updated Icon component to support new icon types.
3. Refactored Toaster component to allow more customization.
4. Completely reworked useToast hook for better type safety and flexibility.
5. Added ToastContent component for consistent toast styling.
6. Updated toast variants to include 'warning' and changed 'error' to 'destructive'.
7. Improved toast rendering with custom icons and animations.
8. Enhanced type definitions for better developer experience.

These changes provide a more robust and customizable toast system, improving the overall user interface and developer experience when working with notifications in the application.

* refactor(Toaster): remove infinite duration from InfoToast story

Removes the infinite duration setting from the InfoToast story to align it with default behavior and improve consistency across toast examples. This change allows the InfoToast to use the default duration, making it more representative of typical usage scenarios.

* feat: make title semibold when description exists

* chore: remove comment from code and use it in the PR

* chore: adjust icon color

* fix: pagination dots to be centered