Add skip composition prop for portal based comps (#37)
* feat: dont use transform to center modal content on big screens

We do that to avoid creating a new containing block for fixed positioning. This will allow any dropdowns or popovers to overflow the modal when we do not use portals.

* feat(ui): add skipComposition option to Dropdown, Popover, Select, and Tooltip components

Introduces a new `skipComposition` prop to Dropdown, Popover, Select, and Tooltip components, allowing users to bypass the default Portal wrapper. This enhancement provides more flexibility in component composition, especially useful in scenarios where Portal behavior is not desired or causes issues with certain layouts. The change also exports Portal components for each, giving developers direct access to these primitives when needed.

* feat(storybook): add ShowcaseSkipComposition story to demonstrate nested components behavior

This new story showcases the behavior of nested components (Modal, Popover, and Dropdown) with and without using portals. It illustrates how skipping composition affects the rendering and functionality of these nested UI elements, providing developers with a clear example of when to use the skipComposition prop for better component integration in complex layouts.

* feat(ShowcaseSkipComposition): add z-index support for nested components

Introduces a new prop 'withZIndexes' to demonstrate proper stacking of nested components using z-index. This change allows for correct rendering of overlapping elements (Modal, Popover, Dropdown) when using portals. Two new stories are added to showcase the difference between using and not using correct z-indexes, providing a clear example of how to handle layering in complex UI compositions.

* chore: bump version

---------

Co-authored-by: Alex Bratsos <alexisbratsos@gmail.com>